% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{batch_estimate_density}
\alias{batch_estimate_density}
\title{Batch Density Estimation}
\usage{
batch_estimate_density(
  training_data,
  query_points,
  class_index,
  bandwidth = 0.1
)
}
\arguments{
\item{training_data}{Training data in polar coordinates}

\item{query_points}{Matrix of query points}

\item{class_index}{Class to estimate density for}

\item{bandwidth}{Bandwidth parameter}
}
\value{
Vector of density estimates
}
\description{
Batch Density Estimation
}
\keyword{internal}
