% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{classify_dm}
\alias{classify_dm}
\title{Quantum Matrix Classifier - Main Interface}
\usage{
classify_dm(x, D_tilde, bandwidth = DM_DEFAULTS$bandwidth)
}
\description{
Core functions for quantum machine learning classification
}
