% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{compute_breaks_list}
\alias{compute_breaks_list}
\title{Compute breaks list for numeric columns}
\usage{
compute_breaks_list(data, n_breaks = 3)
}
\description{
Compute breaks list for numeric columns
}
\keyword{internal}
