% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{compute_quantum_density_matrix}
\alias{compute_quantum_density_matrix}
\title{Compute Quantum Density Matrix}
\usage{
compute_quantum_density_matrix(train_data)
}
\arguments{
\item{train_data}{List of training data by class}
}
\value{
Quantum density matrix with proper normalization
}
\description{
Creates the quantum density matrix (rho_d) from training data
}
\keyword{internal}
