% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{compute_trace}
\alias{compute_trace}
\title{Compute Matrix Trace}
\usage{
compute_trace(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
Numeric value of the trace
}
\description{
Calculates the trace (sum of diagonal elements) of a matrix
}
\keyword{internal}
