% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{convert_to_polar_coordinates}
\alias{convert_to_polar_coordinates}
\title{Convert to Polar Coordinates}
\usage{
convert_to_polar_coordinates(data_list)
}
\arguments{
\item{data_list}{List of coordinate data}
}
\value{
List of data in polar coordinates
}
\description{
Converts Cartesian coordinates to polar coordinates
}
\keyword{internal}
