% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\name{create_model_summary}
\alias{create_model_summary}
\title{Create Model Summary Statistics}
\usage{
create_model_summary(data, model_params)
}
\arguments{
\item{data}{Training data by class}

\item{model_params}{Model parameters}
}
\value{
List of summary statistics
}
\description{
Create Model Summary Statistics
}
\keyword{internal}
