% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{create_quantum_matrix_format}
\alias{create_quantum_matrix_format}
\title{Create Quantum Matrix Representation}
\usage{
create_quantum_matrix_format(data, test_var)
}
\arguments{
\item{data}{Input data with encoded variables}

\item{test_var}{Name of test indicator variable}
}
\value{
Data frame in quantum matrix format
}
\description{
Converts encoded data to final quantum matrix format
}
\keyword{internal}
