% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{dm}
\alias{dm}
\title{Legacy Quantum Matrix Classification Function}
\usage{
dm(...)
}
\arguments{
\item{...}{All parameters (ignored)}
}
\value{
Error message directing to new interface
}
\description{
This function is deprecated. Use dm_fit() and predict() instead.
}
\keyword{internal}
