% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\docType{data}
\name{DM_DEFAULTS}
\alias{DM_DEFAULTS}
\title{Utility Functions for Quantum Matrix Classification}
\format{
An object of class \code{list} of length 9.
}
\usage{
DM_DEFAULTS
}
\description{
Common helper functions, constants, and configurations
}
\keyword{datasets}
