% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{encode_logical_variables}
\alias{encode_logical_variables}
\title{Encode Logical Variables}
\usage{
encode_logical_variables(data, test_var)
}
\arguments{
\item{data}{Input data frame}

\item{test_var}{Name of test indicator variable to preserve}
}
\value{
Data frame with logical variables encoded
}
\description{
Converts logical variables to one-hot encoding
}
\keyword{internal}
