% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{encode_target_variable}
\alias{encode_target_variable}
\title{Encode Target Variable}
\usage{
encode_target_variable(data, objective_var)
}
\arguments{
\item{data}{Input data frame}

\item{objective_var}{Name of objective variable}
}
\value{
Data frame with objective variable encoded as class indicators
}
\description{
Converts objective variable to one-hot class indicators
}
\keyword{internal}
