% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{estimate_class_density}
\alias{estimate_class_density}
\title{Advanced Kernel Density Estimation}
\usage{
estimate_class_density(
  training_data,
  query_point,
  class_index,
  bandwidth = 0.1,
  kernel_type = "quantum"
)
}
\arguments{
\item{training_data}{Training data in polar coordinates}

\item{query_point}{Point to evaluate density at}

\item{class_index}{Class to estimate density for}

\item{bandwidth}{Bandwidth parameter (default = 0.1)}

\item{kernel_type}{Type of kernel}
}
\value{
Estimated probability density
}
\description{
Advanced Kernel Density Estimation
}
\keyword{internal}
