% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{f_hat_h}
\alias{f_hat_h}
\title{Probability Density Estimation}
\usage{
f_hat_h(D_tilde, x, clase, h_window = 0.1)
}
\arguments{
\item{D_tilde}{Training data in polar coordinates (list by class)}

\item{x}{Point to evaluate density at (vector of coordinates)}

\item{clase}{Class index to estimate density for}

\item{h_window}{Bandwidth parameter for kernel (default = 0.1)}
}
\value{
Estimated probability density value
}
\description{
Probability Density Estimation
}
