% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{get_C_tilde_polar}
\alias{get_C_tilde_polar}
\title{Transform Data to Polar Coordinates}
\usage{
get_C_tilde_polar(
  Data,
  objective_var = NULL,
  n_breaks = 3,
  verbose = 0,
  test_var = "test",
  zero_limit = 4
)
}
\arguments{
\item{Data}{Input data with test/train indicators}

\item{objective_var}{Name of objective variable}

\item{n_breaks}{Number of breaks for discretization}

\item{verbose}{Verbosity level (0 = silent, >2 = detailed output)}

\item{test_var}{Name of test indicator variable}

\item{zero_limit}{Precision limit for SVD operations}
}
\value{
List with training and test data in polar coordinates
}
\description{
Main workflow function that transforms quantum matrix data to polar coordinates:
\enumerate{
\item Creates quantum density matrix using get_rho_d()
\item Performs SVD decomposition
\item Projects data onto eigenvector basis
\item Converts to polar coordinate system
\item Validates distribution differences between classes
}
}
