% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{get_u_coordinates}
\alias{get_u_coordinates}
\title{Get Coordinates in Eigenvector Basis}
\usage{
get_u_coordinates(X_tilde, U)
}
\arguments{
\item{X_tilde}{Normalized quantum state vector (row matrix)}

\item{U}{Matrix of eigenvectors from SVD (columns = eigenvectors)}
}
\value{
Coordinates in the eigenvector basis
}
\description{
Projects normalized vector onto eigenvector basis from SVD
}
\keyword{internal}
