% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\name{is_effectively_zero}
\alias{is_effectively_zero}
\title{Check Numeric Tolerance}
\usage{
is_effectively_zero(x, tolerance = DM_DEFAULTS$eigenvalue_tolerance)
}
\arguments{
\item{x}{Numeric value to check}

\item{tolerance}{Tolerance level}
}
\value{
Logical indicating if value is within tolerance of zero
}
\description{
Check Numeric Tolerance
}
\keyword{internal}
