% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{K}
\alias{K}
\title{Kernel Functions for Quantum Matrix Classification}
\usage{
K(..., r = 2)
}
\arguments{
\item{...}{Coordinate values to evaluate kernel at}

\item{r}{Kernel parameter (default = 2)}
}
\value{
Kernel value at the specified coordinates
}
\description{
Mathematical kernel functions and probability estimation for quantum classification
}
