% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n_levels.R
\name{plot_n_levels}
\alias{plot_n_levels}
\title{Visualize Levels of Categorical Variables}
\usage{
plot_n_levels(dataset, lan = "en")
}
\arguments{
\item{dataset}{A data frame containing factor variables}

\item{lan}{Language for labels ("en" or "es")}
}
\value{
A ggplot object
}
\description{
Bar plot showing number of levels for each factor variable
}
