% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{predict.dm_fit}
\alias{predict.dm_fit}
\title{Predict method for quantum matrix classifier}
\usage{
\method{predict}{dm_fit}(object, newdata, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A fitted model object of class "dm_fit"}

\item{newdata}{A data frame containing new observations to predict}

\item{type}{Type of prediction: "class" for predicted classes, "prob" for class probabilities}

\item{...}{Additional arguments (currently unused)}
}
\value{
For type="class": factor vector of predicted classes.
For type="prob": tibble with class probabilities (columns = classes)
}
\description{
Make predictions using a fitted quantum matrix classifier.
Follows tidymodels conventions: uses stored model parameters without refitting.
}
\examples{
model <- dm_fit(Species ~ ., data = iris)
predict(model, iris)
predict(model, iris, type = "prob")

}
