% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{prepare_dm_data}
\alias{prepare_dm_data}
\title{Prepare Data for Quantum Matrix Processing}
\usage{
prepare_dm_data(data, objective_var, n_breaks = 3, test_var = "test")
}
\arguments{
\item{data}{Input data frame}

\item{objective_var}{Name of objective variable}

\item{n_breaks}{Number of breaks for continuous variables}

\item{test_var}{Name of test indicator variable}
}
\value{
List with processed data components
}
\description{
Converts and preprocesses data for quantum matrix operations
}
\keyword{internal}
