% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{print.dm_fit}
\alias{print.dm_fit}
\title{Print method for quantum matrix classifier}
\usage{
\method{print}{dm_fit}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class "dm_fit"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print a summary of the fitted quantum matrix classifier
}
