% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml-package.R
\docType{package}
\name{qml-package}
\alias{qml}
\alias{qml-package}
\title{qml: Quantum Machine Learning for Classification}
\description{
Quantum machine learning classification using Density Matrix (DM) algorithm.
Main functions: dm_fit(), predict(), prepare_split_data().
}
\section{Quick Start}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Basic usage
model <- dm_fit(Species ~ ., data = iris, n_breaks = 3)
predictions <- predict(model, newdata = iris)

# With rsample
library(rsample)
split <- initial_split(iris, strata = Species)
prepared_data <- prepare_split_data(split, "Species")
model <- dm_fit(Species ~ ., data = prepared_data)
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RodGal-2020/qml}
  \item Report bugs at \url{https://github.com/RodGal-2020/qml/issues}
}

}
\author{
\strong{Maintainer}: Xopre Rodríguez Gallego \email{rodgal.2020@example.com} (\href{https://orcid.org/0000-0002-5825-2724}{ORCID})

}
\keyword{internal}
