% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\name{round_precision}
\alias{round_precision}
\title{Round to Significant Digits}
\usage{
round_precision(x, digits = DM_DEFAULTS$zero_limit)
}
\arguments{
\item{x}{Numeric value or vector}

\item{digits}{Number of significant digits}
}
\value{
Rounded value
}
\description{
Round to Significant Digits
}
\keyword{internal}
