% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe Division}
\usage{
safe_divide(numerator, denominator, replacement = 0)
}
\arguments{
\item{numerator}{Numerator value}

\item{denominator}{Denominator value}

\item{replacement}{Value to return if denominator is zero}
}
\value{
Result of division or replacement value
}
\description{
Safe Division
}
\keyword{internal}
