% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{separate_data_by_class}
\alias{separate_data_by_class}
\title{Separate Data by Class}
\usage{
separate_data_by_class(data, objective_var, n_classes)
}
\arguments{
\item{data}{Quantum matrix formatted data}

\item{objective_var}{Name of objective variable}

\item{n_classes}{Number of classes}
}
\value{
List of data frames, one per class
}
\description{
Splits quantum matrix data by class labels
}
\keyword{internal}
