% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{split_train_test_data}
\alias{split_train_test_data}
\title{Split Train and Test Data}
\usage{
split_train_test_data(class_data, test_var)
}
\arguments{
\item{class_data}{List of data frames by class}

\item{test_var}{Name of test indicator variable}
}
\value{
List with train and test data components
}
\description{
Separates class data into training and test sets
}
\keyword{internal}
