% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{summary.dm_fit}
\alias{summary.dm_fit}
\title{Summary method for quantum matrix classifier}
\usage{
\method{summary}{dm_fit}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "dm_fit"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Provide a detailed summary of the fitted quantum matrix classifier
}
