% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{test_distribution_differences}
\alias{test_distribution_differences}
\title{Test for Distribution Differences}
\usage{
test_distribution_differences(polar_data)
}
\arguments{
\item{polar_data}{List of polar coordinate data by class}
}
\value{
List with test results
}
\description{
Uses Fasano-Franceschini test to check if class distributions differ
}
\keyword{internal}
