% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{transform_to_coordinates}
\alias{transform_to_coordinates}
\title{Transform Data to Coordinate System}
\usage{
transform_to_coordinates(data_list, U, rank_rho)
}
\arguments{
\item{data_list}{List of class-specific data}

\item{U}{Matrix of eigenvectors}

\item{rank_rho}{Rank of the quantum matrix}
}
\value{
List of transformed data
}
\description{
Applies coordinate transformation using SVD eigenvectors
}
\keyword{internal}
