% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{validate_dm_input}
\alias{validate_dm_input}
\title{Validate Input Data for Quantum Matrix Classification}
\usage{
validate_dm_input(data, objective_var, n_breaks = 3)
}
\arguments{
\item{data}{Input data frame or tibble}

\item{objective_var}{Name of the objective variable}

\item{n_breaks}{Number of breaks for discretization}
}
\value{
TRUE if validation passes, otherwise throws an error
}
\description{
Validates that the input data meets requirements for the quantum matrix classifier
}
\keyword{internal}
