% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_utils.R
\name{validate_probabilities}
\alias{validate_probabilities}
\title{Validate Probability Vector}
\usage{
validate_probabilities(probs, tolerance = DM_DEFAULTS$trace_tolerance)
}
\arguments{
\item{probs}{Numeric vector of probabilities}

\item{tolerance}{Tolerance for sum being 1}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate Probability Vector
}
\keyword{internal}
