% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{validate_quantum_matrix}
\alias{validate_quantum_matrix}
\title{Validate Quantum Density Matrix Properties}
\usage{
validate_quantum_matrix(rho_d)
}
\arguments{
\item{rho_d}{Input matrix to validate}
}
\value{
TRUE if valid, otherwise throws warning
}
\description{
Checks if a matrix satisfies quantum density matrix properties
}
\keyword{internal}
