% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{discretize_continuous_features}
\alias{discretize_continuous_features}
\title{Quantum Matrix Operations for Classification}
\usage{
discretize_continuous_features(data, n_breaks = 3, breaks_list = NULL)
}
\arguments{
\item{data}{Input data frame}

\item{n_breaks}{Number of breaks for discretization}
}
\value{
Data frame with continuous variables discretized
}
\description{
Functions for creating and manipulating quantum density matrices

Converts continuous variables to factors using cut() function
}
\keyword{internal}
