% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{dm_fit}
\alias{dm_fit}
\title{Quantum Matrix Method for Classification - Model Fitting}
\usage{
dm_fit(formula, data, n_breaks = 3, verbose = 0, ...)
}
\arguments{
\item{formula}{A formula specifying the model (e.g., y ~ .)}

\item{data}{A data frame containing the training data}

\item{n_breaks}{Number of breaks for discretizing continuous variables}

\item{verbose}{Verbosity level (0 = silent, 1+ = more verbose)}

\item{...}{Additional parameters}
}
\value{
An object of class "dm_fit" containing the fitted model
}
\description{
Fits a quantum matrix classifier using the DM model.
Returns a fitted model object that can be used with predict().

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Fit a quantum matrix classifier
model <- dm_fit(Species ~ ., data = iris)
predictions <- predict(model, iris)

}
