% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_classifier.R
\name{empty_prediction_result}
\alias{empty_prediction_result}
\title{Create Empty Prediction Result}
\usage{
empty_prediction_result(type, classes, return_coords = FALSE)
}
\arguments{
\item{type}{Prediction type}

\item{classes}{Model classes}

\item{return_coords}{Whether to include coordinates}
}
\value{
Empty result in correct format
}
\description{
Create Empty Prediction Result
}
\keyword{internal}
