% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{encode_categorical_variables}
\alias{encode_categorical_variables}
\title{Encode Categorical Variables}
\usage{
encode_categorical_variables(data, objective_var, test_var)
}
\arguments{
\item{data}{Input data frame}

\item{objective_var}{Name of objective variable}

\item{test_var}{Name of test indicator variable}
}
\value{
Data frame with factor variables encoded
}
\description{
Converts factor variables to one-hot encoding
}
\keyword{internal}
