% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transforms.R
\name{get_preprocessing_params}
\alias{get_preprocessing_params}
\title{Coordinate Transformation Functions for Quantum Matrix Classification}
\usage{
get_preprocessing_params(
  Data,
  objective_var = NULL,
  n_breaks = 3,
  verbose = 0,
  test_var = "test",
  zero_limit = 4,
  bandwidth = DM_DEFAULTS$bandwidth
)
}
\arguments{
\item{Data}{Training data}

\item{objective_var}{Name of objective variable}

\item{n_breaks}{Number of breaks for discretization}

\item{verbose}{Verbosity level}

\item{test_var}{Name of test indicator variable}

\item{zero_limit}{Precision limit for SVD operations}
}
\value{
List with preprocessing parameters for predictions
}
\description{
Functions for transforming data to polar coordinates and other coordinate systems

Extracts and stores preprocessing parameters needed for making predictions.
This follows tidymodels pattern of separating fit and predict phases.
}
\keyword{internal}
