% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_matrix.R
\name{get_rho_d}
\alias{get_rho_d}
\title{Create Quantum Density Matrix from Data}
\usage{
get_rho_d(
  Data,
  objective_var = NULL,
  n_breaks = 3,
  verbose = 0,
  test_var = "test"
)
}
\arguments{
\item{Data}{Input data with test/train indicator}

\item{objective_var}{Name of objective variable}

\item{n_breaks}{Number of breaks for discretization}

\item{verbose}{Verbosity level (0 = silent, >2 = detailed output)}

\item{test_var}{Name of test indicator variable}
}
\value{
List containing rho_d matrix, training data, and test data
}
\description{
Main function that orchestrates the complete quantum matrix creation process
}
